+++ src/controllers/FcdsController.ts
@@ -1,6 +1,5 @@
 import { inject, injectable } from 'inversify';
 import express, { Request, Response } from 'express';
-import { AuthUtils } from '../services/AuthUtils';
 import FCD from '../models/FCD';
 import Settings from '../types/Settings';
 
@@ -10,7 +9,7 @@ class FcdsController {
 
   router: express.Application;
 
-  constructor(@inject(AuthUtils) private readonly authUtils: AuthUtils) {
+  constructor() {
     this.router = express().get('/', this.index).get('/:chainId', this.index);
   }