+++ test/e2e/apiKeys/getApiKeys.e2e.ts
@@ -1,146 +1,70 @@
-import axios from 'axios';
+import 'reflect-metadata';
+import { Container } from 'inversify';
+import { Application } from 'express';
 import { loadTestEnv } from '../../helpers';
-import chai, { expect } from 'chai';
-import chaiAsPromised from 'chai-as-promised';
-import User from '../../../src/models/User';
-import Project from '../../../src/models/Project';
-import ApiKey from '../../../src/models/ApiKey';
-import mongoose from 'mongoose';
+import { setupDatabase, teardownDatabase } from '../../helpers/databaseHelpers';
+import { getContainer } from '../../../src/lib/getContainer';
+import Server from '../../../src/lib/Server';
+import { after } from 'mocha';
+import request from 'supertest';
+import { expect } from 'chai';
+import Project, { IProject } from '../../../src/models/Project';
+import ApiKey, { IApiKey } from '../../../src/models/ApiKey';
+import { setupAuthHarness, TestAuthHarness } from '../../helpers/authHelpers';
+import { projectFactory } from '../../mocks/factories/projectFactory';
+import { apiKeyFactory } from '../../mocks/factories/apiKeyFactory';
 
-chai.use(chaiAsPromised);
-
-describe('Get API keys', () => {
-  const config = loadTestEnv();
-  const appAxios = axios.create({ baseURL: config.APP_URL });
-  let accessToken: string;
-  let firstProjectId: string;
-  let secondProjectId: string;
+describe('/api-keys', async () => {
+  let container: Container;
+  let app: Application;
 
   before(async () => {
-    await mongoose.connect(config.MONGODB_URL, { useNewUrlParser: true, useUnifiedTopology: true });
-  });
-
-  beforeEach(async () => {
-    await User.deleteMany({});
+    loadTestEnv();
+    await setupDatabase();
     await Project.deleteMany({});
     await ApiKey.deleteMany({});
-
-    await appAxios.post('/users', {
-      email: 'example@example.com',
-      password: 'valid_password',
-    });
-
-    const response = await appAxios.post('/auth', {
-      email: 'example@example.com',
-      password: 'valid_password',
-    });
-
-    accessToken = response.data.token;
-
-    const firstProjectResponse = await appAxios.post(
-      '/projects',
-      { name: 'First project' },
-      { headers: { authorization: `Bearer ${accessToken}` } }
-    );
-    firstProjectId = firstProjectResponse.data._id;
-
-    const secondProjectResponse = await appAxios.post(
-      '/projects',
-      { name: 'Second project' },
-      { headers: { authorization: `Bearer ${accessToken}` } }
-    );
-    secondProjectId = secondProjectResponse.data._id;
+    container = getContainer();
+    app = container.get(Server).app;
   });
 
   after(async () => {
-    await User.deleteMany({});
     await Project.deleteMany({});
     await ApiKey.deleteMany({});
-
-    await mongoose.connection.close();
+    await teardownDatabase();
   });
 
-  it('responds with 403 if no access token was provided', async () => {
-    await expect(
-      appAxios.get('/api-keys', {
-        headers: {
-          authorization: '',
-        },
-      })
-    ).to.be.rejected.then((error) => {
-      expect(error.response.status).to.be.eq(403);
-    });
-  });
-
-  it('returns keys for both projects if projectId is not specified', async () => {
-    // a key for the first project
-    await appAxios.post(
-      '/api-keys',
-      {
-        projectId: firstProjectId,
-      },
-      { headers: { authorization: accessToken } }
-    );
-
-    // a key for the second project
-    await appAxios.post(
-      '/api-keys',
-      {
-        projectId: secondProjectId,
-      },
-      { headers: { authorization: accessToken } }
-    );
-
-    const response = await appAxios.get('/api-keys', {
-      headers: {
-        authorization: accessToken,
-      },
+  describe('GET /api-keys', async () => {
+    describe('when the user is not authenticated', async () => {
+      it('responds with HTTP 401 Unauthorized', async () => {
+        const res = await request(app).get('/api-keys');
+        expect(res.status).to.eq(401);
+      });
     });
 
-    expect(response.status).to.be.eq(200);
+    describe('when the user is authenticated', async () => {
+      let project: IProject;
+      let apiKey: IApiKey;
+      let authHarness: TestAuthHarness;
 
-    const apiKeys: Record<string, unknown>[] = response.data;
+      beforeEach(async () => {
+        authHarness = await setupAuthHarness();
+        project = await Project.create({ ...projectFactory.build(), ownerId: 'USER_ID', ownerType: 'User' });
+        apiKey = await ApiKey.create({ ...apiKeyFactory.build(), projectId: project.id });
+      });
 
-    const apiKeyForFirstProject = apiKeys.find((apiKey) => apiKey.projectId === firstProjectId);
-    expect(apiKeyForFirstProject).to.be.an('object', 'Not found an API key for the first project');
-    const apiKeyForSecondProject = apiKeys.find((apiKey) => apiKey.projectId === secondProjectId);
-    expect(apiKeyForSecondProject).to.be.an('object', 'Not found an API key for the second project');
-  });
+      afterEach(async () => {
+        await authHarness.jwksMock.stop();
+      });
 
-  it('returns keys only for the specified project when projectId is provided', async () => {
-    // a key for the first project
-    await appAxios.post(
-      '/api-keys',
-      {
-        projectId: firstProjectId,
-      },
-      { headers: { authorization: accessToken } }
-    );
+      it('responds with the API Keys', async () => {
+        const res = await request(app).get('/api-keys').set('Authorization', `Bearer ${authHarness.accessToken}`);
 
-    // a key for the second project
-    await appAxios.post(
-      '/api-keys',
-      {
-        projectId: secondProjectId,
-      },
-      { headers: { authorization: accessToken } }
-    );
-
-    const response = await appAxios.get('/api-keys', {
-      params: {
-        projectId: firstProjectId,
-      },
-      headers: {
-        authorization: accessToken,
-      },
-    });
-
-    expect(response.status).to.be.eq(200);
-
-    const apiKeys: Record<string, unknown>[] = response.data;
-
-    apiKeys.forEach((apiKey) => {
-      expect(apiKey.projectId).to.be.eq(firstProjectId);
+        expect(res.status).to.eq(200);
+        expect(res.body.length).to.eq(1);
+        expect(res.body[0]._id).to.eq(apiKey._id);
+        expect(res.body[0].projectId).to.eq(apiKey.projectId);
+        expect(res.body[0].key).to.eq(apiKey.key);
+      });
     });
   });
 });