+++ CHANGELOG.md
@@ -4,7 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [4.6.0] - 2021-11-17
+### Added
+- Auth0 Authentication
+- Profile API
+- Metrics middleware
+
+### Removed
+- DB-based authentication
+- Auth routes
+- WalletAuth
 
 ## [4.5.0] - 2021-11-17
 ### Added